+++ theme.js
@@ -1,0 +1,33 @@
+import { createMuiTheme } from '@material-ui/core/styles';
+
+// Create a theme instance.
+const theme = createMuiTheme({
+  palette: {
+    primary: {
+      // main: '#fec838',
+      main: '#F15F3A',
+      contrastText: '#FFFFFF',
+    },
+    secondary: {
+      main: '#FFFFFF',
+      dark: '#2E2E2E',
+      contrastText: '#333333',
+      text: '#525252',
+    },
+  },
+  props: {
+    MuiButtonBase: {
+      disableRipple: true,
+    },
+  },
+  typography: {
+    fontFamily: 'Open Sans',
+  },
+  breakpoints: {
+    values: {
+      sm: 768,
+    },
+  },
+});
+
+export default theme;