+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [Unreleased](https://github.com/nathanab/beacon/compare/master...HEAD)
 
 ## [v1.2.0](https://github.com/nathanab/beacon/compare/v1.1.4...v1.2.0) - TBD
 - Add admin mode with date builder